requirements.txt
appnope==0.1.3
asttokens==2.2.1
attrs==23.1.0
backcall==0.2.0
bcrypt==4.0.1
beautifulsoup4==4.12.2
blinker==1.6.2
click==8.1.3
decorator==5.1.1
dnspython==2.3.0
email-validator==2.0.0.post2
executing==1.2.0
Flask==2.3.2
Flask-Bcrypt==1.0.1
Flask-DebugToolbar==0.13.1
flask-expects-json==1.7.0
Flask-JWT==0.3.2
Flask-WTF==1.1.1
greenlet==2.0.2
gunicorn==20.1.0
idna==3.4
ipython==8.14.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.38
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.15.1
PyJWT==1.4.2
pyrsistent==0.19.3
python-dotenv==1.0.0
six==1.16.0
soupsieve==2.4.1
SQLAlchemy==2.0.15
stack-data==0.6.2
traitlets==5.9.0
typing_extensions==4.6.3
wcwidth==0.2.6
Werkzeug==2.3.4
WTForms==3.0.1
